refactor: create sections/agents with shared AgentFilters and AgentCard
Consolidate duplicate filter logic from AgentsTable and
AgentsNavigationPage into a shared useAgentFilters hook in
sections/agents/AgentFilters.tsx. Move AgentCard from sections/cards
to sections/agents for co-location.
